# Speedup `Node::is_greater_than` by avoiding `alloca`